+++ typescript/solver/solvers/eco/filler.ts
@@ -17,8 +17,7 @@ import { logDebug, logError, logGreen } from "../../logger.js";
 import { IntentCreatedEventObject } from "../../typechain/eco/contracts/IntentSource.js";
 import { Erc20__factory } from "../../typechain/factories/contracts/Erc20__factory.js";
 import { EcoAdapter__factory } from "../../typechain/factories/eco/contracts/EcoAdapter__factory.js";
-import { HyperProver__factory } from "../../typechain/factories/eco/contracts/HyperProver__factory.js";
-import { IntentSource__factory } from "../../typechain/factories/eco/contracts/IntentSource__factory.js";
+import { withdrawRewards } from "./utils.js";
 
 type IntentData = { [targetAddress: string]: BigNumber };
 
@@ -30,7 +29,7 @@ export const create = () => {
     ORIGIN_SETTLER_CHAIN_ID,
   } = setup();
 
-  return async function onChain(intent: IntentCreatedEventObject) {
+  return async function eco(intent: IntentCreatedEventObject) {
     logGreen("Received Intent:", intent._hash);
 
     const result = await prepareIntent(
@@ -209,51 +208,4 @@ async function fill(
   }
 
   logDebug("Fulfilled intent on", _chainId, "with data", _data);
-}
-
-async function withdrawRewards(
-  intent: IntentCreatedEventObject,
-  originSettlerAddress: string,
-  originChainId: string,
-  multiProvider: MultiProvider,
-) {
-  const { _hash, _prover } = intent;
-
-  logGreen("Waiting for `IntentProven` event on origin chain");
-  const signer = multiProvider.getSigner(originChainId);
-  const claimantAddress = await signer.getAddress();
-  const prover = HyperProver__factory.connect(_prover, signer);
-
-  await new Promise((resolve) =>
-    prover.once(
-      prover.filters.IntentProven(_hash, claimantAddress),
-      async () => {
-        logDebug("Intent proven:", _hash);
-
-        logGreen("About to claim rewards");
-        const settler = IntentSource__factory.connect(
-          originSettlerAddress,
-          signer,
-        );
-        const tx = await settler.withdrawRewards(_hash);
-
-        const receipt = await tx.wait();
-
-        const baseUrl =
-          multiProvider.getChainMetadata(originChainId).blockExplorers?.[0].url;
-
-        if (baseUrl) {
-          logGreen(
-            `Withdraw Rewards Tx: ${baseUrl}/tx/${receipt.transactionHash}`,
-          );
-        } else {
-          logGreen("Withdraw Rewards Tx:", receipt.transactionHash);
-        }
-
-        logDebug("Reward withdrawn on", originChainId, "for intent", _hash);
-
-        resolve(_hash);
-      },
-    ),
-  );
 }
